refactor: s3 저장 로직 리펙토링
