+++ matchingService/src/main/java/com/nbcamp/gamematching/matchingservice/member/domain/AmazonS3ResourceStorage.java
@@ -1,4 +1,4 @@
-package com.nbcamp.gamematching.matchingservice.member.controller;
+package com.nbcamp.gamematching.matchingservice.member.domain;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
