+++ matchingService/src/main/resources/templates/profile/my-profile.html
@@ -35,7 +35,7 @@
                         <div id="profile-content-box" class="panel-body">
                             <div class="row">
                                 <div id="panel-body__img-box">
-                                    <img src="${imageUrl}"/>
+                                    <img src="https://game-matching.s3.ap-northeast-2.amazonaws.com/${imageUrl}"/>
                                 </div>
                                 
                                 <table class="table table-user-information profile-box__info">
